requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
supabase==1.0.4
python-dotenv==1.0.0
